Worked on base64! :)

+++ src/_04_Base64_Decoder/Base64DecoderTester.java
@@ -33,13 +33,14 @@ public class Base64DecoderTester {
 		answer[0] = (byte) 255;
 		answer[1] = (byte) 255;
 		answer[2] = (byte) 255;
+		
 		check = Base64Decoder.convert4CharsTo24Bits("////");
-		// 111.111 111.111 111.111 111.111		// 1111.1111 1111.1111 1111.1111
+		// 111.111 111.111 111.111 111.111		// 1111.1111 1111.1111 1111.1111 
 		assertArrayEquals(answer, check);
 
-		answer[0] = (byte) 7;
-		answer[1] = (byte) 13;
-		answer[2] = (byte) 126;
+		answer[0] =  7;
+		answer[1] =  13;
+		answer[2] =  126;
 		check = Base64Decoder.convert4CharsTo24Bits("Bw1+");
 		// 000.001 110.000 110.101 111.110
 		// 0000.0111 0000.1101. 0111.1110
